Break down bnae compat configs, fix styling
