+++ optionals/bnae_falkor_compat_ace/CfgMagazines.hpp
@@ -1,0 +1,8 @@
+class CfgMagazines {
+    class CA_Magazine;
+
+    // AtragMx GunList ".300WM Berger OTM"
+    class 10Rnd_300WM_Magazine: CA_Magazine {
+        initSpeed=800; // muzzle velocity according with the barrel length, default 990
+    };
+};
