+++ optionals/bnae_mk1_compat_ace/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\tac\addons\bnae_mk1_compat_ace
