+++ optionals/bnae_mk1_compat_ace/CfgMagazines.hpp
@@ -1,0 +1,8 @@
+class CfgMagazines {
+    class CA_Magazine;
+
+    class 10Rnd_303_Magazine: CA_Magazine {
+       ammo="B_303_Ball";
+       initSpeed=763; // muzzle velocity according with the barrel length
+    };
+};
