+++ optionals/bnae_mk1_compat_ace/README.md
@@ -1,0 +1,5 @@
+# About
+
+Add [ACE3](https://github.com/acemod/ACE3) compatibility for [bnae Mk1](https://forums.bohemia.net/forums/topic/196809-project-infinite-v10/).
+
+### Authors
