+++ optionals/bnae_mk1_compat_ace/script_component.hpp
@@ -1,0 +1,17 @@
+#define COMPONENT bnae_mk1_compat_ace
+#define COMPONENT_BEAUTIFIED BNAE Mk1 ACE3 Compat
+#include "\x\tac\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_BNAE_MK1_COMPAT_ACE
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_BNAE_MK1_COMPAT_ACE
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_BNAE_MK1_COMPAT_ACE
+#endif
+
+#include "\x\tac\addons\main\script_macros.hpp"
