+++ optionals/bnae_rk95_compat_ace/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\tac\addons\bnae_rk95_compat_ace
