+++ optionals/bnae_trg42_compat_ace/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\tac\addons\bnae_trg42_compat_ace
