+++ optionals/bnae_trg42_compat_ace/README.md
@@ -1,0 +1,5 @@
+# About
+
+Add [ACE3](https://github.com/acemod/ACE3) compatibility for [bnae trg42](https://forums.bohemia.net/forums/topic/196809-project-infinite-v10/).
+
+### Authors
